+++ src/routes/root_route.py
@@ -3,7 +3,7 @@ from flask import Response, Blueprint
 from json import dumps as to_json
 from semantic_search_engine.constants import SYNC_INTERVAL_SHELVE, LAST_SYNC_TIME_SHELVE, CHROMA_SHELVE
 from routes.login_decorator import login_required
-from semantic_search_engine.mattermost.fetch_mm_data import is_sync_inprogress
+from semantic_search_engine.mattermost.sync_posts import is_sync_inprogress
 from semantic_search_engine.shelves import retrieve_one
 from . import mattermost
